[IMP] account_accountant: prevent account change of deferred line
When opening the list view of journal items, a user can use multi-select
to change the account of one or more lines. In the case where we have
lines that are already deferred, changing the account would break the
balance deferred by the already existing deferral moves.

In order to prevent this, we don't allow changing the account of a line
that is already deferred.

task-3516660

Part-of: odoo/enterprise#48022