[IMP] helpdesk_timesheet: trigger compute_project_id with ticket changed
Before this commit, when the user selects a ticket for a timesheet
via the timesheet header in the grid view for instance, the
project set on the timesheet is not recomputed to have the one
set on the hepdesk team of the ticket linked.

This commit ensures the project is recomputed when the ticket is set
on a timesheet to be sure the project be the one set on the helpdesk
team linked to the ticket set.

part of task-2276015
task-3542388

Part-of: odoo/enterprise#48523