[FIX] helpdesk_timesheet: recheck if the project set is one in helpdesk
Before this commit, when the user changes the project in the `project_id`
field inside the timesheet header component and that project is not linked
to a helpdesk team than the `helpdesk_ticket_id` field should be
replaced by `task_id` one.

This commit replaces `hasHelpdeskProject` attribute into a getter
to call it dynamically when it is needed.

task-3542388

Part-of: odoo/enterprise#48523