[IMP] account_*: show grouped journal items for deferrals
*account_accountant,account_reports

Smart buttons allow navigating between invoices/bills and their
deferral moves. When clicking a smart button, a list of moves is shown.
In order to audit deferrals and their original invoices, you have to
open every move separately, which makes it very inconvenient.

In order to easy auditability and get an instant overview of what
happens with the deferrals, we now show a list of journal items grouped
by move.

task-3516660

closes odoo/enterprise#48022

Signed-off-by: William André (wan) <wan@odoo.com>