[REF] point_of_sale: remove AlertPopup
The AlertPopup component is identical to the ErrorPopup component,
aside from the fact that the ErrorPopup also plays a sound. We
thus add a new prop to the ErrorPopup that allows the consumer
to decide to have the sound played or not and replace all occurences of the
AlertPopup with ErrorPopup.

closes odoo/enterprise#47987

Task: 3522638
Related: odoo/odoo#136700
Signed-off-by: Robin Heinz (rhe) <rhe@odoo.com>