[FIX] spreadsheet_edition: wrong datasource in pivot side panel
When the side panel of a pivot is opened while the side panel of
another pivot is already open, the datasource of the first pivot is
used instead of the datasource of the second pivot. This can lead to
tracebacks, or to wrong data being displayed.

Task: 3463289
Part-of: odoo/enterprise#48283