[IMP] account_followup: more descriptive studio UI
The title of the followup report was made editable in studio in a
previous commit.
But the UI still looks a bit rough in studio:
Below the title there are just multiple strings saying
"Expression", since the actual content depends on the result of some
functions called in the template.
This commit replaces the "Expression" strings with some more
descriptive strings.

task-3446724

closes odoo/enterprise#47633

Signed-off-by: Quentin De Paoli <qdp@odoo.com>