[FIX] account_reports: use column group options to audit carryover lines
When doing a comparison between multiple periods with values carried over to them, auditing the carryover values (through the dedicated button in the popup, in debug mode) always opened the latest period.

This was due to the fact the main options of the report were directly used, instead of the ones corresponding to the column group owning the value.

closes odoo/enterprise#48030

X-original-commit: a2ebae8894649856a8e3aad75a77bc28e930ef35
Related: odoo/odoo#136868
Signed-off-by: John Laterre (jol) <jol@odoo.com>
Signed-off-by: Olivier Colson (oco) <oco@odoo.com>