[FIX] web_studio: element creation for insertion
When the OdooEditor is mounted in an iframe, creating an element using
the top document's `createElement` method produces a different result
than creating one with the Editor's document. While the former is an
instance of the top document's global `Node`, the latter is not (it is
an instance of the iframe's global `Node` instead).

This commit fixes an undesired failed comparison using the `instanceof`
operator in the editor's "insert" command, due to not using the Editor's
document when creating a new element for insertion.

task-3521152

closes odoo/enterprise#48136

Related: odoo/odoo#136597
Signed-off-by: David Monjoie (dmo) <dmo@odoo.com>