[IMP] hr_appraisal: adapt views to new employee fields
In the community counterpart of this PR, we change the employee custom fields
to stop using always 'hr.employee.public' as the relation, but dynamically
decide which relation it should use depending on the group that the current
user has. Also, we allow forcing a relation by passing it as an option in the
view. So, this commit removes the relation of the fields that would benefit
from having this feature of showing different records depending on the user
group.

task-3524305

closes odoo/enterprise#48008

Related: odoo/odoo#136786
Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>