[FIX] account_reports: warnings z-index
There was an issue with the z-index of 'warnings' and 'btn_group'.

To make the 'warnings' be in front of '.o_view_nocontent', we forced it to have a 'z-index' of 999. This worked, but it made a very strange side effect. Sometimes, we couldn't click on the dropdown items on the control panel of a report.

The issue was that 'btn-group' forced a 'z-index' of 2 so what actually happened is that the dropdown was behind (even if not visibly) 'warnings' so the click was actually being done on it and not the dropdown items.

To fix that we move '.o_view_nocontent' back with the 'z-index' of -1.

closes odoo/enterprise#47969

Forward-port-of: #47909
Signed-off-by: John Laterre (jol) <jol@odoo.com>
Signed-off-by: Olivier Colson (oco) <oco@odoo.com>