[FIX] hr_contract_salary: search for correct bank account in configurator
In salary configurator, when employee signs the contract, we are looking
for existing bank account. If bank account exists, we set it for the employee.

Here, to find the existing bank account, the search domain uses only
'acc_number'. This can be an issue, in case two employees, A and B have
the same bank account.

Because for employee B search might return the bank account belonging to
Employee A (as they have the same acc_number).
While setting this bank account for employee B, we check whether
address_home_id of the current employee is the same as
bank_account.partner_id. And if it is not the case, we untrust
the account.

To solve this issue, we need to search for existing bank account
not only by acc_number, but also by partner_id.

closes odoo/enterprise#48033

X-original-commit: cc1658c434a75d8aa105b14e3c2161f83bcafe80
Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>