[IMP] {website_}appointment: move anytime appt with recurring
Purpose
=======
Move the anytime appointment types from the Invitations
menu to the Appointments tab to be displayed with the
recurring appointment types.

Specifications
==============
The reason is that the anytime appointment types are
just recurring ones with slots defined from 7AM to 7PM
on Monday to Friday.

In the appointment type search view:
- add filters on the recurring and anytime category
- add a filter on is_published
- add a group by category
- set the recurring appointment type filter as default.

In the appointment type kanban view:
- change the appointment type location by the
appointment category.

Task-3109254

Part-of: odoo/enterprise#43298