[IMP] account_reports: introduce sum_children keyword in aggregations
It is now possible for aggregations to be defined with the formula 'sum_children'. An expression using this formula will take all the expressions with the same label, belonging to child lines of the line this expression belongs to, and will sum their values. This is done as an improvement in stable, as this feature is needed for a cleaner upgrade of custom-made tax reports from 15.0 to 16.0, since the former tax report engine used to behave like that by default (no formula meant you had to sum the children).

Task 3499260

closes odoo/enterprise#48019

Related: odoo/odoo#136812
Signed-off-by: John Laterre (jol) <jol@odoo.com>