[FIX] knowledge: fix template selection
When a user chooses the same template from the template gallery twice,
the preview disappears from the screen. This commit aims to resolve this
issue.

Steps to reproduce the issue:
1. Open the template gallery of Knowledge
2. Click on a template to preview it
3. Click on the same template again
=> The preview disappears from the screen

TO BE: When the user selects the same template twice, the scroll
offset of the preview should be set to 0.

To reset the scroll offset of the preview when the user selects a
template, we follow these steps to avoid any flickering:

1. We hide the preview
2. We set the scroll offset of the preview back to 0
3. We load the new record
4. When the component is patched and the res id of the record changes,
we display the preview to the user.

When the user clicks on the same template twice, the step 4 is not
executed because the res id of the record does not change (see: the
dependencies of the `useEffect`). Hence, the preview remains in the
hidden state. This commit will ensure that the preview will be hidden
only if the user selects another template.

Reference: https://github.com/odoo/enterprise/pull/47451

task-3525333

closes odoo/enterprise#48848

X-original-commit: 009bfc9b9c45983f2c2986a3dca77df0d9d6c6df
Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>