[FIX] spreadsheet_edition: `SpreadsheetControlPanel` missing a component
`SpreadsheetControlPanel` is spawning a `ControlPanel` in order to set
some of its `slots`. It then calls a template `spreadsheet_edition.Breadcrumbs`
which itself depends on `Dropdown`.
With the current architecture, the template is called in the context of
`SpreadsheetControlPanel`, which does not have `Dropdown`defined in its
static components.

The issue was partially solved by fixing the components of
`DocumentsSpreadsheetControlPanel` but the issue could arise for actions
that use `SpreadsheetControlPanel` directly.

closes odoo/enterprise#48233

Task: /
X-original-commit: 9139585682b810c52f6d3bb03ac82fee01f6d2a1
Signed-off-by: Lucas Lefèvre (lul) <lul@odoo.com>
Signed-off-by: Rémi Rahir (rar) <rar@odoo.com>