[FIX] account_accountant: sync. bank rec. widget with list view
The bank reconciliation kanban view is currently not updated after changes to
transactions in the list view. This can lead to wrong (outdated)
information to be shown on the right side in the bank reconciliation kanban view.
This further allows creating wrong reconcilation journal entries with the "Validate" button
(I.e. after a change to the "amount" of a transaction in the list view
is not reflected on the right side of the bank reconciliation kanban view).

The test tour was extended to test for this case.

To reproduce:
  1. Go to bank reconciliation (in kanban view)
  2. Select some transaction kanban card that is not yet reconciled
  3. Switch to the tree view
  4. Modify the "amount" field of the transaction that was selected in step 2.
     This corresponds to a change in the database.
  5. Switch back to kanban view
  6. The right part of the UI (under the "Validate" button) is not updated:
     I.e. the top line (liquidity line) has a wrong debit / credit.
  7. Validate
  8. The journal entry created due the validation in step 7 uses the
     non-updated / wrong values and is thus wrong.

task-3504471

closes odoo/enterprise#47930

Signed-off-by: Habib Ayob (ayh) <ayh@odoo.com>