[FIX] pos_settle_due: allow opening POS without accounting rights
If a POS has a different currency than the company currency, when
opening the POS without accounting access rights, it would fail with
an key error when trying to get the total_due on partners.

To reproduce:

- Create a new POS journal with a non-company currency
- Create a new POS using that journal
- Remove accounting access rights from the POS user
- Open the POS as that user

It would throw an key error when trying to get the total_due.

This commit fixes it by skipping the total_due computation if the
user does not have accounting access rights.

opw-3512997

closes odoo/enterprise#48073

Signed-off-by: Robin Heinz (rhe) <rhe@odoo.com>