[FIX] account_reports: Remove the bold weight of some report line.
Problem
---------
Some report lines where bolded even though they had no children. This
should not be the case. The case where a line is bold even though it has
no sublines occurs when all the sublines are hidden because they equals
to 0. Thus, they exist but are not shown.

Objective
---------
Make the report line not bold when its sublines are hidden because they
eqeual 0.

Solution
---------
Add a condition on the `unfolded` report line class so that it is not
given to lines that have no visible children.

task-3508362

closes odoo/enterprise#48485

X-original-commit: 55e12dc7655dad4460545d51ae5169e7cf42edf6
Signed-off-by: John Laterre (jol) <jol@odoo.com>
Signed-off-by: Antoine Boonen (aboo) <aboo@odoo.com>