[FIX] web_studio: report editor: avoid superfluous render
Before this commit, the component ReportEditorXMLEditor was subscribed
to the "_isInEdition" key in the reportEditorModel. This caused glitchy (but eventually harmless)
renderings

This was due to the reactivity, which gets the value before setting it. In our case,
getting the value via a getter, actually subscribed the reader to the underlying private key (in the getter's code).

After this commit, the key "_isInEdition" is written via a function and not a getter. This stops the subscription to the key of the component setting the key. Hence, there is no superfluous
render because of this particular key.

Part-of: odoo/enterprise#48385