[IMP] industry_fsm_sale: block arrows behaviour in edit mode
This commit's purpose is to block the up and down arrows from
increasing/decreasing the product quantity of an fsm task when the user
is in 'input' mode.
Indeed, when a user pused one of the arrows, it triggers and increase or
a decrease of 1 of the current value. The issue is that an update is
directly triggered each time the key is pushed, even if the displayed
value has not yet been updated. This can lead to the user doing
hundreds of operation in a short time if he maintains the key pushed
down, and the ui will take quite some time to getting up to date.

task-3489879
https://www.odoo.com/web#id=3489879&menu_id=4720&cids=1&action=333&active_id=4105&model=project.task&view_type=form

closes odoo/enterprise#48171

Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>