[FIX] account_winbooks_import: fix forgotten onboarding step
Code related to step was forgotten in the onboarding's refactor in
Task-3025136 which removed a function called at the end of the
winbooks import therefore making it fail.

Note however that the step itself was not shown on the dashboard panel
as its template was removed in d09cf75.

Task-3524102

Part-of: odoo/enterprise#48163