[IMP] account_reports,*: replace options 'multi_company' / 'single_company' with 'companies'
*: account_asset,account_disallowed_expenses,l10n_es_reports,l10n_eu_oss_reports,l10n_lu_reports,l10n_mx_xml_polizas,l10n_nl_reports_sbr,l10n_pe_reports,l10n_sg_reports,mrp_account_enterprise,mrp_workorder_hr_account

Main changes
1. Report options:
   Instead of storing either a 'multi_company' or 'single_company' key just store a 'companies' key.
2. Simplify the template 'pdf_export_filters' from 'account_reports'
3. Adapt calls to function '_get_query_currency_table' of model 'res.currency'.
   Due to 1. this function was refactored (see community PR).

on 1.
-----
The new 'companies' key is used to store all the active companies (even if there is only one).

This is the main change. The other changes were possible (2.) or
needed (3.) due to this change.

All the modules depending on it had to be adapted accordingly.

on 2.
-----
The printing of the company names was simplified.
A comma was added as separator between the company names.
The render value 'report_company_name' that was only used to print the company name in the
"single company" case was removed.

on 3.
-----
Due to change 1 the '_get_query_currency_table' function of the
res.currency model (added in module 'account' from the community repository) was refactored:
It does not use the report options as argument anymore (see community PR).
A helper function '_get_query_currency_table' was introduced in model 'account.report' that just takes
the report options as a parameter and calls '_get_query_currency_table' with the proper arguments.
All calls to '_get_query_currency_table' were adapted

task-3484368

Part-of: odoo/enterprise#47735