[FIX] l10n_ec_edi: Show correct message error in Authorization state
The authorization status could be "EN PROCESO" which means the document was processed by SRI but it didn't return anything.
The normal process is blocked because an error message is displayed instead of a simple warning. The document should be sent on the next request and it should be handled by the cron job.
It was detected through that many clients generate a lot of invoices per minute and the SRI leaves some documents in authorization status "EN PROCESO", that is where the error was detected.
To solve it, return the authorization state in _l10n_ec_get_authorization_status method in order to validate if the edi document state  is "EN PROCESO" and show the correct warning.

closes odoo/enterprise#48168

Signed-off-by: Josse Colpaert <jco@odoo.com>