[FIX] documents: check registry in domain getter
The link_to_record wizard was updated to fix the domain getter
in [1].
However the upgrade of the module will now often fail
because the view is checked when some models are not loaded in
the model registry of the environment.

We fix this by simply checking that it is actually in the env as that's
a sensible behavior.

[1]: f6b0eddc10fd65d65dc68b2d3333dd1850f48515

task-3487592

closes odoo/enterprise#48223

X-original-commit: 6391311552d50d86012b532d071768b9e13065ab
Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>