[FIX] account_online_synchronization: Fix multi companies issue
When the user has to select a journal for its bank accounts,
the wizard made a suggestion of the journal. This suggestion is
based partially on the currency.

How to reproduce:
Have 2 companies (one based in USA and one in BE for example),
each company should have a bank journal with no operation and with
the default currency set (USD for USA and EUR for BE).
Set the BE company as default company and tries to make a bank
synchronization. Select a USD bank account and synchronize.
The wizard will suggest you the USD bank journal (from USA company)
and it will raise a multi companies issue.

The bug is caused because we do not pass the selected company in the
context, then the controller will select all available companies for
this user. The fix adds the selected company in the context.

Ideally, it will better to make a check in python that the selected
company is the account.online.account company. But it was deciced that
this fix is sufficient as the bug is only from 16.0 to 16.2 (excluded).

task-3272565
opw-3247755

closes odoo/enterprise#48207

Signed-off-by: John Laterre (jol) <jol@odoo.com>