[FIX] timer: TimerStartField shouldn't use onWillUpdateProps
Since the new model (PR: odoo/odoo#114024), updating a record no longer
triggers a deep render and therefore no longer triggers the onWillUpdateProps
for Field components.

The goal of this commit is to adapt the usage of onWillUpdateProps
in Field composents in order to fix the bugs introduced by the RelationalModel.

This commit will also ensure that the setInterval used in TimerStartField
is always clear.

closes odoo/enterprise#47991

Signed-off-by: Aaron Bohy (aab) <aab@odoo.com>