[FIX] account_avatax: fix taxes computation in multi-currency
Steps to reproduce:
- Install Accounting, Sales, Contacts & AvaTax
- Setup AvaTax in Accounting settings
- Activate CAD currency
- Create a CAD pricelist
- Create a fiscal position:
  * Use AvaTax API: checked
  * Detect Automatically: checked
  * Country: Canada
- Create a Canadian contact: (e.g. Contact X)
  * Street: 2 Rue des Jardins
  * City: Quebec
  * State: Quebec (CA)
  * ZIP: G1R459
  * Country: Canada
- Create a product with "[FR] Shipping only common carrier - fob destination (backward compatibility)"
as Avatax Category. (e.g. Product X)
- Create a SO:
  * Customer: Contact X
  * Pricelist: CAD pricelist
  * Order Lines:
    - product: Product X
    - Unit Price 100.00
- Save SO
- Click on "COMPUTE TAXES USING AVATAX"
=> 2 taxes are added ("CANADA GST/TPS [CA] (5.0000%)" and "QUEBEC QST/TVQ [QC] (9.9750%)").
The 2nd one is exempted. So the total taxes are 5.00.
- Confirm SO
- Deliver the product if needed
- Create an invoice from SO (Regular invoice)
- Click on "COMPUTE TAXES USING AVATAX"

1st issue:
The total taxes are 14.98, which are different from the taxes computed in SO (i.e. 5.00).
The exempted tax is not exempted in the total taxes of the invoice.

2nd issue:
When checking the tax lines in "Journal Items" tab, we can see that the amounts of the
debit/credit columns are not consistent.
For "CANADA GST/TPS [CA] (5.0000%)" tax, the credit has the same value than the absolute
value of "Amount in Currency", when the currency rate should be applied.
For "QUEBEC QST/TVQ [QC] (9.9750 %)" tax (that should be exempted), debit and credit are
0, but "Amount in Currency" is not.

Cause:
When checking if the tax should be manually fixed, we are comparing tax line balance with
the amount provided by AvaTax.
However, these 2 amounts are not always in the same currency.
Also, when fixing the tax amount manually, amount_currency field is not updated and
debit/credit fields are updated with the values coming from AvaTax without conversion to
the company currency.

opw-3439742

closes odoo/enterprise#49364

X-original-commit: b532fe0583517700e709e118f3e1a8d68aea9949
Signed-off-by: William André (wan) <wan@odoo.com>
Signed-off-by: Anh Thao Pham (pta) <pta@odoo.com>