[FIX] account_reports: Unfolding behavior correction
Description of the issue/feature this commit addresses:

When using the Danish localization and going in the Profit and Loss report
or going in the Lithuanian Balance Sheet report, if all lines are folded and
then unfolded, some subsections do not appear as they are the child of
a section which has not been defined as foldable and therefore unfoldable.
These sections should be shown when the report is unfolded.

This behavior is believed to be present in other reports for other
localizations but these two were the examples that were used for the task.

Desired behavior after the commit is merged :

This commit modifies the unfolding process so a new check is done.
When a line is unfolded, a check is done on each of its children. If they
are not foldable AND they have at least one child, they are unfolded too.

Adding this fix, all subsections are now available at all time. Notably in
the Danish P&L and in the Lithuanian BS but most probably in other places too.

task-3462024

X-original-commit: 1f192ac6bc55005e6669005fb67761be9303b7dc
Part-of: odoo/enterprise#48636