[FIX] industry_fsm_stock: avoid computing qty delivered when it’s set
This fix is necessary for this PR: https://github.com/odoo/odoo/pull/135865

Because the "test_multiple_fsm_task" test fails at this line:
https://github.com/odoo/enterprise/blob/16.0/industry_fsm_stock/tests/test_fsm_stock.py#L739-L740

When the "action_fsm_validate" function is called, the "qty_delivered"
is computed:

https://github.com/odoo/enterprise/blob/0025b1295319f69feb70d5cb5ec80acf92d76ec9/industry_fsm_sale/models/project_task.py#L236

https://github.com/odoo/enterprise/blob/21ba5cca5107f2ac16e956d669338ecb44622d00/industry_fsm_stock/models/project_task.py#L32

https://github.com/odoo/enterprise/blob/21ba5cca5107f2ac16e956d669338ecb44622d00/industry_fsm_stock/models/project_task.py#L41

However, in the "_compute_qty_delivered" function, since the moves are
in the "confirmed" status, the "qty_delivered" is set to 0:

https://github.com/odoo/odoo/blob/8a3284685856b6980e3b620abdbbb3e2c79f57c4/addons/sale_stock/models/sale_order.py#L215

But then, this quantity is modified with "so_line.product_uom_qty":

https://github.com/odoo/enterprise/blob/21ba5cca5107f2ac16e956d669338ecb44622d00/industry_fsm_stock/models/project_task.py#L65-L66

However, since move lines are created  afterward:
https://github.com/odoo/enterprise/blob/21ba5cca5107f2ac16e956d669338ecb44622d00/industry_fsm_stock/models/project_task.py#L67

The "qty_delivered" field must be recomputed because it now depends on
"qty_done."

But since the moves remain in the "confirmed" status, the
"qty_delivered" is set to 0, whereas it should remain at 2

Solution:

Set the "qty_delivered" with "so_line.product_uom_qty" after the
creation of move lines. This way, the "compute_qty_delivered" will
not be triggered.

OPW-3504138

closes odoo/enterprise#48547

X-original-commit: f371a540ebcd449f7aefef89aab0f83106a97e35
Related: odoo/odoo#137869
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>