[FIX] web_studio: better reactivity for approvals
This commit fixes two issues:
- Have a button that has two approvals: one with a domain, the other without
with the pager change record from one that has the two approvals, to the one that has only one
(or vice versa). Before this change, the approvals thumbnails did not update

- Same case, but save one record that did have only one approval to make it have two (or vice versa)
Before this commit, the approvals thumbnails did not update.

This change set corrects those two use cases, and technically adapts the approval_hook to the new relational model, with which the record is itself reactive, meaning that we don't always go through willUpdateProps, rather,
we use useRecordObserver.

opw-3498555
opw-3382099

closes odoo/enterprise#48278

Signed-off-by: Luca Vitali (luvi) <luvi@odoo.com>