[IMP] mrp_workorder: shop floor barcode improvements
This commit makes some improvements to the handling of barcode commands
in the shop floor display:
1. Close the employee selection popup when scanning a badge
2. Don't allow scanning of WO's not visible in the view
3. New behavior for the close MO/WO and pause commands:
Always pick the first record in the view that the current employee is
working on. If the current employee is not working on any records in the
view, just pick the first record.

Task: 3456827
Part-of: odoo/enterprise#48379