[FIX] account_online_synchronization: Only get aml with a journal_id set
Sometimes, we have customers having some account.move.line without
journal_id. It's not a normal usecase! It could be related to
migration or something else but it causes an issue with the bank
sync wizard when we're trying to get all journals having amls.
This commit adds a check on the read_group to only get amls
having a journal_id set.

opw-3524416

closes odoo/enterprise#48282

Signed-off-by: John Laterre (jol) <jol@odoo.com>