[FIX] l10n_be_reports: tax reports: don't rely on modifying options to report failed checks
In 16.3, in case some checks failed on the tax report, the XML export wizard displayed a banner informing the user some checks had not been successful.

From 16.4 on, it is not allowed to modify the options dict while generating the lines, and doing so has no effect anymore (since options are generated from a distinct RPC call). Because of that, the banner did show up anymore in the Belgian export wizard.

We cannot restore the behavior of the banner (as it would mean we have to recompute the report when generating the options, which is of course not acceptable). Instead, we make the failed checks more visible on the report, by using a red banner and the standard warnings mechanism. The wizard does not show anything anymore, but it is assumed the user cannot open the export wizard without having first noticed the big red banner on the report.

closes odoo/enterprise#48326

X-original-commit: 2be6526e57a05eda13c8a5aa7e873f5ffac071b4
Signed-off-by: John Laterre (jol) <jol@odoo.com>
Signed-off-by: Olivier Colson (oco) <oco@odoo.com>