[FIX] account_reports: aged receivable/payable: don't format foreign currency amounts with domestic currency
To reproduce:

1) Create an invoice of 100 EUR in a company in USD
2) Open the Aged Payable/Receivable (depending on the type of invoice created in 1); works with both)
3) Unfold the partner you created the invoice for

===> The line displays an amount currency of 100$ instead of 100€. The currency displayed on the line is EUR, though (which is correct).

This bug was due to the fact the amount_currency expression of those reports uses the monetary figure_type. This hence formatted the amount with the main currency of the company. This was of course wrong ; to solve it, we introduce the _currency_* prefix on expression labels, a new syntax to be used in custom engines to handle cases like this one. When an expression if prefixed _currency_xxx, it means its value corresponds to the currency id the amount of expression labelled xxx (on the same report line) is expressed in.

This bug was already fixed in 16.4 by https://github.com/odoo/enterprise/pull/43885 (we probably mistakenly thought this bug was introduced by the refactoring made in that branch). This commit is a backport of what was made there.

closes odoo/enterprise#48303

Signed-off-by: John Laterre (jol) <jol@odoo.com>