[FIX] appointment: enable AppointmentInviteCopyClose
Since the new Relational Model, if the <a> in AppointmentInviteCopyClose
is disabled, it is not possible to enable it when you resolve alerts.

Why:
---
Before the new Relational Mode, the entire view was re-rendered each
time the record was updated. AppointmentInviteCopyClose was therefore
rendered and the callback in useEffect was called. This is no longer the case,
only the Components that observe the data modified on the record are rendered.

Solution:
---------
We're going to use the dynamicProps.readonly which will be recalculated
each time a change occurs in its condition. So we'll pass the readonly
attribute to the <widget> with the condition the combination of all the
conditions that can display an alert and cause the <a> to be disabled.

closes odoo/enterprise#48309

Related: odoo/odoo#137414
Signed-off-by: Aaron Bohy (aab) <aab@odoo.com>