[FIX] sale_amazon: better matching for the states codes from amazon
Amazon sometimes allows user to type their state and/or state code instead of forcing the customers to choose in a list, and then send the information as typed. As such, before this commit, even if we already considered the case where the state code is in fact the complete state name, the search was case-sensitive. We now use the operator to match the state, as long as it at least doesn't have any typo.

opw-3140019

closes odoo/enterprise#48735

X-original-commit: 84e9db6d0cb77ed88c6c21e9cc3ab008d9661f6f
Signed-off-by: Antoine Vandevenne (anv) <anv@odoo.com>