[FIX] knowledge: update behaviorState validator onWillUpdateProps
In [1], a new strategy was implemented to detect whether a `Behavior` is
obsolete for the current state of the `HtmlField` it is mounted for.

However, when switching record, the `MutationObserver` is triggering the
mounting process for any `Behavior` in the new record just before the
`onPatched` hook of the `useEffect` which is supposed to update the `validator`
object which represents the current state of the `HtmlField`, meaning that
those behaviors mounting process is started with obsolete values from the get
go, resulting in an early abortion and a re-mounting just after.

This can be avoided by updating the `validator` object `onWillUpdateProps` with
the new props, which occurs before the `MutationObserver` can detect the `Owl`
patch.

[1]: https://github.com/odoo/enterprise/commit/ee8771ebe9f3789e22172d644581bd310815615e

task-3172056

Part-of: odoo/enterprise#43488