[IMP] web_studio: granular behaviour for footer and header
Before this commit, when reseting the report the footer and header were always included.
The available fields when editing the header or footer were no different from the ones
in the content of the report.

After this commit, when reseting the report, a confirmation dialog with the choice to
include the footer and header is shown.

Also, when adding a field with the field popover in the header/footer, the fields available are the ones of the object `company` on the
report, that is, the fields of `res.company`.

task-id-3457404

Part-of: odoo/enterprise#47089