[IMP] web_studio: Save and Discard on report editor
Before this commit, when editing a report, if the save operation failed, the
changes that the user made were lost.

After this commit, the changes are not lost if the save operation fails. A button
"Discard" is added. Now, the user has to act to discard the changes.

task-id-3457404

Part-of: odoo/enterprise#47089