[IMP] web_studio: report editor loading error resilience
Before this commit, when the initial RPC to load the report editor crashed, Studio redirected
to the actionEditor.

After this commit, it simply puts an hardcoded html in the iframe. This way, the report editor
is still usable even though the report is not valid. It makes it easier to go to the XMLEditor to
fix the views by hand.

Part-of: odoo/enterprise#48385