[IMP] mrp_workorder: shop floor reload measure step after correction
When making a correction to a measurement type quality check in the shop
floor, the quality check line on the workorder was not updated to show
the new saved value. This commit makes sure the view is updated when
closing any action windows after validating quality checks.

Task: 3456827 (nr 25)
Part-of: odoo/enterprise#48379