[IMP] account_assets: unfold all
Before this commit when printing the pdf, all the lines were unfolded. We now
try to make the pdf as close as what the user see in the interface. So it means
that if the lines are folded in the report we print them as such.

Task: 3437257
Part-of: odoo/enterprise#45792