[FIX] account_accountant: fix attachments not showing in journal items
After the changes made in odoo/odoo/pull/114024, a related field is not fetched if invisible. This causes the mimetype in main_attachment_ids field to not be fetched in the Journal Items list view therefore no attachments are shown in the preview.

Solution: remove the fields from the arch xml entirely and manually add them as an active field in the config.

task-3531478

closes odoo/enterprise#48372

Signed-off-by: Habib Ayob (ayh) <ayh@odoo.com>