[REF] knowledge: remove knowledge_embedded_view_framework key from action context
When inserting an embedded view, we currently store a key named
`knowledge_embedded_view_framework` in the context of the action
responsible for loading the embedded view. This key is used to
distinguish the embedded views from other views, enabling us to
selectively hide specific interface elements, such as the
"Insert view in article" button for embedded views.

Instead of retaining this key in the action's context, we can make
use of the embedded view ID assigned when mounting the embedded view
in the DOM. The embedded view ID is unique to embedded views and
will always be set for them. Consequently, we can use this ID to
distinguish between views, eliminating the need to store an
additional key in the action's context.

This commit will remove the `knowledge_embedded_view_framework` key
from the context of the action responsible for loading embedded views.
Additionally, it will update the code logic to rely on the embedded
view ID rather than the now-unnecessary key.

task-3334717

Part-of: odoo/enterprise#48691