[IMP] account_reports: Introduce new tag selection syntax into account_codes engine
Some countries don't have a legally fixed chart of accounts, and hence cannot use the 'account_codes' engine in their Balance Sheet and P&L. Instead they have to rely on account.account.tag objects associated with the accounts of the database, and target them in the reports using the 'domain' engine.

However, the 'domain' engine lacks the feature achieved with the C and D suffixes in 'account_codes'. Indeed, in the latter, writing '123D' will consider all the accounts prefixed 123 if their balance on the period is positive (C means negative). This is not doable with a domain.

The financial reports engine, used pre-16.0, used to support that case with sum_if_pos_groupby and sum_if_neg_groupby. We mistakenly considered those syntaxes would not be required with tags, which is why we're making this commit in stable.

This commit introduces a new tag(...) selector, usable with the account_codes engine just like any regular prefix selector. It can take an xmlid (for data files) or an account.account.tag id (for custos in the UI; if the tag was created manually) as parameter, and can be used together with account codes in the same formula. It also supports C and D suffixes and prefix exclusion. It is hence possible to combine it with the pre-existing syntax; for example:

tag(account.some_tag)\(42)D + 1234C + tag(42) - 456\(4560)

Task 3524409

closes odoo/enterprise#49011

X-original-commit: a858bebefd7f8075d8a74436bb2d41fcea3731b7
Signed-off-by: John Laterre (jol) <jol@odoo.com>
Signed-off-by: Olivier Colson (oco) <oco@odoo.com>