[IMP] {website_}appointment{_hr}: add non permanent appt types
Purpose
=======
Allow the creation of appointment types for
punctual periods.

Specification
=============
Add optionals start and end datetimes on the appointment
type form to limit the slot generation between 2 specified
datetimes.
If the start and end datetimes are specified, hide the
maximum number of days to schedule.

The appointment types of category 'website' can now be
either 'recurring' by default or 'punctual'.
Renaming the 'website' category of the appointment type
throughout all modules.

Task-3109254

Part-of: odoo/enterprise#43298