[MOV] quality{,_control}: move to quality_control
Both functions use the `measure_on` field of `quality.point` which is
defined in the `quality_control` module.
This didn't raise any issue as both functions are only used in
`quality_control` and `quality_mrp` (which requires `quality_control`).

Moved them to `quality_control` as this is where they should have
probably been in the first place.

closes odoo/enterprise#48428

Signed-off-by: Tiffany Chang (tic) <tic@odoo.com>