[IMP] web_studio: copy the initial view when editing it
Before this commit, the report editor altered the view in place, with no means to see what had been
modified.

After this commit, the original view is saved as a new ir.ui.view.
Moreover, the base diff tool for views is integrated in the reportEditor to allow
for easily spotting differences in the new XML versus the old one.

This commit should improve the upgrade process.

task-id-3531025

closes odoo/enterprise#48385

Signed-off-by: Bruno Boi (boi) <boi@odoo.com>